pyproject.toml
[tool.poetry]
name = "review-visualizer"
version = "0.1.0"
description = ""
authors = ["Rochan Madhusudhana <rochana.hm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.1"
pandas = "^2.1.2"
streamlit = "^1.28.0"
st-pages = "^0.4.5"
plotly-express = "^0.4.1"
sqlalchemy = "^2.0.22"
streamlit-searchbox = "^0.1.6"
mysql-connector-python = "^8.2.0"
tqdm = "^4.66.1"
prisma = "^0.11.0"
networkx = "^3.2.1"
httpcore = { extras = ["asyncio"], version = "^1.0.2" }

[tool.poetry.group.dev.dependencies]
pyclean = "^2.7.5"
black = "^23.10.1"
pytest = "^7.4.3"
poethepoet = "^0.24.1"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"
isort = "^5.12.0"

[tool.poe.tasks]
dev = ["generate", "start-dev"]
start-dev = "streamlit run review_visualizer/app/main.py"
generate = "prisma generate"
clean = "pyclean ."
blk = "black ."
fmt = "isort ."
format = ["blk", "fmt"]
test = "pytest"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
